+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "alloglot" extension will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adhere's to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- Add tags-based "completions", "go to definitions", and "suggest imports" code action.
+
 ## [2.1.1] - 2024-02-07
 
 - Fix bug in document formatter that prevented text edits if the formatter logs to stderr.